requirements.txt
# requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
google-cloud-firestore>=2.16.0,<3
requests>=2.31.0,<3
httpx>=0.27.0,<1
numpy>=1.26.0,<2
Pillow>=10.3.0,<11
pytz>=2024.1
pydantic==2.7.4
firebase-admin>=6.4.0,<7
anyio>=4.0,<5
slowapi>=0.1.9,<1
cryptography>=42.0.0,<44

ftfy>=6.2.0,<7
regex>=2024.4
tqdm>=4.66.0,<5

# Testing
pytest>=8.0,<9
pytest-asyncio>=0.23,<1

